WMP001-040 Aktualizacja TransactionMapper i TransactionMapperTest

+++ src/main/java/com/kodilla/walletmanager/dto/TransactionDto.java
@@ -62,7 +62,7 @@ public class TransactionDto {
             return this;
         }
 
-        public TransactionDtoBuilder amount(long amount){
+        public TransactionDtoBuilder amount(double amount){
             this.amount = amount;
             return this;
         }
